+++ src/main/java/duke/commands/AddTaskCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.components.Storage;
+import duke.components.TaskList;
+import duke.components.Ui;
+import duke.tasks.Task;
+
 import java.io.IOException;
 
 /**
