+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.components.Storage;
+import duke.components.TaskList;
+import duke.components.Ui;
+
 /**
  * Abstract Command class that can be executed to perform a command
  */
