+++ src/main/java/duke/commands/DeleteTaskCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.components.Storage;
+import duke.components.TaskList;
+import duke.components.Ui;
+import duke.tasks.Task;
+
 /**
  * Command used to delete a Task from the TaskList
  */
