+++ src/main/java/duke/tasks/Event.java
@@ -1,3 +1,7 @@
+package duke.tasks;
+
+import duke.components.Parser;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
