+++ src/main/java/duke/tasks/Task.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 import java.time.LocalDate;
 
 /**
